LPD-77985 [FE] Move common logic to History component
