+++ modules/apps/site/site-cmp-site-initializer/src/main/resources/META-INF/resources/js/components/task/TaskInfoSummary.tsx
@@ -4,6 +4,7 @@
  */
 
 import Label from '@clayui/label';
+import {DateRenderer} from '@liferay/frontend-data-set-web';
 import {AssigneeValue} from '@liferay/object-dynamic-data-mapping-form-field-type';
 import {displayErrorToast} from '@liferay/site-cms-site-initializer';
 import React from 'react';
@@ -14,13 +15,11 @@ import {
 	displayStateSuccessToast,
 } from '../../utils/toastUtil';
 import CustomAssignee from '../CustomAssignee';
+import {UPDATE_HISTORY} from '../History';
 import InfoSummary from '../InfoSummary';
 import StateSelector, {State} from '../StateSelector';
 
 import '../AssigneeTrigger.scss';
-import {UPDATE_TASK_HISTORY} from './TaskHistory';
-
-import {DateRenderer} from '@liferay/frontend-data-set-web';
 
 interface TaskInfoSummaryProps {
 	assignTo: AssigneeValue;
@@ -59,7 +58,7 @@ export default function TaskInfoSummary({
 								if (!error) {
 									displayStateSuccessToast();
 
-									Liferay.fire(UPDATE_TASK_HISTORY);
+									Liferay.fire(UPDATE_HISTORY);
 								}
 								else {
 									displayErrorToast(error);
@@ -86,7 +85,7 @@ export default function TaskInfoSummary({
 										(value as AssigneeValue).name
 									);
 
-									Liferay.fire(UPDATE_TASK_HISTORY);
+									Liferay.fire(UPDATE_HISTORY);
 								}
 								else {
 									displayErrorToast(error);
